NIFI-12022 Extract verification logic from JMSConnectionFactoryProvider
